+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [4.23.0] - 2022-03-13
 ### `jsonschema-generator-bom`
 #### Added
@@ -552,6 +554,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
+[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.23.0...HEAD
 [4.23.0]: https://github.com/victools/jsonschema-generator/compare/v4.22.0...v4.23.0
 [4.22.0]: https://github.com/victools/jsonschema-generator/compare/v4.21.0...v4.22.0
 [4.21.0]: https://github.com/victools/jsonschema-generator/compare/v4.20.0...v4.21.0