+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-module-jackson`
+#### Added
+- Support for sub type resolution on properties with `JsonTypeInfo.Id.NONE` override avoiding the default wrapping/modification
 
 ## [4.23.0] - 2022-03-13
 ### `jsonschema-generator-bom`