+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [4.24.0]
+## [Unreleased]
+
+## [4.24.0] - 2022-04-01
 ### `jsonschema-generator`
 #### Fixed
 - When looking-up a matching field or getter, only consider the declared property names and not any overrides
@@ -564,6 +566,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Indicate a number's "exclusiveMaximum" according to `@DecimalMax` or `@Negative`
 
 
+[Unreleased]: https://github.com/victools/jsonschema-generator/compare/v4.24.0...HEAD
 [4.24.0]: https://github.com/victools/jsonschema-generator/compare/v4.23.0...v4.24.0
 [4.23.0]: https://github.com/victools/jsonschema-generator/compare/v4.22.0...v4.23.0
 [4.22.0]: https://github.com/victools/jsonschema-generator/compare/v4.21.0...v4.22.0